requirements.txt
asgiref==3.7.2
Django==4.2.6
Pillow==10.1.0
sqlparse==0.4.4
djangorestframework~=3.14.0
drf_yasg


python-dotenv~=1.0.0